+++ docs/README.md
@@ -42,8 +42,8 @@ The second-largest source of income in Sri Lanka is the apparel industry. While 
 
 ## Links
 
-- [Project Repository](<https://github.com/cepdnaclk/>{{ page.repository-name }}){:target="_blank"}
-- [Project Page](<https://cepdnaclk.github.io/>{{ page.repository-name}}){:target="_blank"}
+- [Project Repository](<https://github.com/cepdnaclk/>{{ page.repository-name }})
+- [Project Page](<https://cepdnaclk.github.io/>{{ page.repository-name}})
 - [Department of Computer Engineering](http://www.ce.pdn.ac.lk/)
 - [University of Peradeniya](https://eng.pdn.ac.lk/)
