+++ src/front-end/src/components/AppBar.tsx
@@ -1,0 +1,44 @@
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import IconButton from '@mui/material/IconButton';
+import Badge from '@mui/material/Badge';
+import MenuIcon from '@mui/icons-material/Menu';
+import NotificationsIcon from '@mui/icons-material/Notifications';
+import { AppBarStyled } from './AppBarProps';
+
+export function AppBar(open: boolean, toggleDrawer: () => void) {
+  return <AppBarStyled position="absolute" open={open}>
+    <Toolbar
+      sx={{
+        pr: '24px', // keep right padding when drawer closed
+      }}
+    >
+      <IconButton
+        edge="start"
+        color="inherit"
+        aria-label="open drawer"
+        onClick={toggleDrawer}
+        sx={{
+          marginRight: '36px',
+          ...(open && { display: 'none' }),
+        }}
+      >
+        <MenuIcon />
+      </IconButton>
+      <Typography
+        component="h1"
+        variant="h6"
+        color="#inherit"
+        noWrap
+        sx={{ flexGrow: 1 }}
+      >
+        AiPaS
+      </Typography>
+      <IconButton color="inherit">
+        <Badge badgeContent={4} color="secondary">
+          <NotificationsIcon />
+        </Badge>
+      </IconButton>
+    </Toolbar>
+  </AppBarStyled>;
+}
