+++ src/front-end/src/components/Sidebar.tsx
@@ -1,0 +1,30 @@
+import Toolbar from '@mui/material/Toolbar';
+import List from '@mui/material/List';
+import Divider from '@mui/material/Divider';
+import IconButton from '@mui/material/IconButton';
+import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
+import { mainListItems, secondaryListItems } from './listItems';
+import { Drawer } from './Drawer';
+
+export default function Sidebar(open: boolean, toggleDrawer: () => void) {
+  return <Drawer variant="permanent" open={open}>
+    <Toolbar
+      sx={{
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'flex-end',
+        px: [1],
+      }}
+    >
+      <IconButton onClick={toggleDrawer}>
+        <ChevronLeftIcon />
+      </IconButton>
+    </Toolbar>
+    <Divider />
+    <List component="nav">
+      {mainListItems}
+      <Divider sx={{ my: 1 }} />
+      {secondaryListItems}
+    </List>
+  </Drawer>;
+}
